+++ app/src/main/java/dev/onecoffeeplz/zamechalka/di/DomainModule.kt
@@ -1,31 +1,31 @@
 package dev.onecoffeeplz.zamechalka.di
 
-import dev.onecoffeeplz.zamechalka.domain.usecase.CreateNoteUseCase
-import dev.onecoffeeplz.zamechalka.domain.usecase.DeleteNoteUseCase
-import dev.onecoffeeplz.zamechalka.domain.usecase.DeleteRecordingUseCase
-import dev.onecoffeeplz.zamechalka.domain.usecase.GetAudioPlaybackPositionUseCase
-import dev.onecoffeeplz.zamechalka.domain.usecase.GetNoteByIdUseCase
-import dev.onecoffeeplz.zamechalka.domain.usecase.GetNotesUseCase
-import dev.onecoffeeplz.zamechalka.domain.usecase.PauseAudioPlayerUseCase
-import dev.onecoffeeplz.zamechalka.domain.usecase.PrepareAudioPlayerUseCase
-import dev.onecoffeeplz.zamechalka.domain.usecase.ReleaseAudioPlayerUseCase
-import dev.onecoffeeplz.zamechalka.domain.usecase.StartAudioPlayerUseCase
-import dev.onecoffeeplz.zamechalka.domain.usecase.StartRecordingUseCase
-import dev.onecoffeeplz.zamechalka.domain.usecase.StopRecordingUseCase
-import dev.onecoffeeplz.zamechalka.domain.usecase.UpdateNoteUseCase
-import dev.onecoffeeplz.zamechalka.domain.usecase.impl.CreateNoteUseCaseImpl
-import dev.onecoffeeplz.zamechalka.domain.usecase.impl.DeleteNoteUseCaseImpl
-import dev.onecoffeeplz.zamechalka.domain.usecase.impl.DeleteRecordingUseCaseImpl
-import dev.onecoffeeplz.zamechalka.domain.usecase.impl.GetAudioPlaybackPositionUseCaseImpl
-import dev.onecoffeeplz.zamechalka.domain.usecase.impl.GetNoteByIdUseCaseImpl
-import dev.onecoffeeplz.zamechalka.domain.usecase.impl.GetNotesUseCaseImpl
-import dev.onecoffeeplz.zamechalka.domain.usecase.impl.PauseAudioPlayerUseCaseImpl
-import dev.onecoffeeplz.zamechalka.domain.usecase.impl.PrepareAudioPlayerUseCaseImpl
-import dev.onecoffeeplz.zamechalka.domain.usecase.impl.ReleaseAudioPlayerUseCaseImpl
-import dev.onecoffeeplz.zamechalka.domain.usecase.impl.StartAudioPlayerUseCaseImpl
-import dev.onecoffeeplz.zamechalka.domain.usecase.impl.StartRecordingUseCaseImpl
-import dev.onecoffeeplz.zamechalka.domain.usecase.impl.StopRecordingUseCaseImpl
-import dev.onecoffeeplz.zamechalka.domain.usecase.impl.UpdateNoteUseCaseImpl
+import dev.onecoffeeplz.zamechalka.domain.usecase.view.CreateNoteUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.view.DeleteNoteUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.recording.DeleteRecordingUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.playback.GetAudioPlaybackPositionUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.view.GetNoteByIdUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.view.GetNotesUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.playback.PauseAudioPlayerUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.playback.PrepareAudioPlayerUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.playback.ReleaseAudioPlayerUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.playback.StartAudioPlayerUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.recording.StartRecordingUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.recording.StopRecordingUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.view.UpdateNoteUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.impl.view.CreateNoteUseCaseImpl
+import dev.onecoffeeplz.zamechalka.domain.usecase.impl.view.DeleteNoteUseCaseImpl
+import dev.onecoffeeplz.zamechalka.domain.usecase.impl.recording.DeleteRecordingUseCaseImpl
+import dev.onecoffeeplz.zamechalka.domain.usecase.impl.playback.GetAudioPlaybackPositionUseCaseImpl
+import dev.onecoffeeplz.zamechalka.domain.usecase.impl.view.GetNoteByIdUseCaseImpl
+import dev.onecoffeeplz.zamechalka.domain.usecase.impl.view.GetNotesUseCaseImpl
+import dev.onecoffeeplz.zamechalka.domain.usecase.impl.playback.PauseAudioPlayerUseCaseImpl
+import dev.onecoffeeplz.zamechalka.domain.usecase.impl.playback.PrepareAudioPlayerUseCaseImpl
+import dev.onecoffeeplz.zamechalka.domain.usecase.impl.playback.ReleaseAudioPlayerUseCaseImpl
+import dev.onecoffeeplz.zamechalka.domain.usecase.impl.playback.StartAudioPlayerUseCaseImpl
+import dev.onecoffeeplz.zamechalka.domain.usecase.impl.recording.StartRecordingUseCaseImpl
+import dev.onecoffeeplz.zamechalka.domain.usecase.impl.recording.StopRecordingUseCaseImpl
+import dev.onecoffeeplz.zamechalka.domain.usecase.impl.view.UpdateNoteUseCaseImpl
 import org.koin.core.module.dsl.bind
 import org.koin.core.module.dsl.factoryOf
 import org.koin.dsl.module
