+++ app/src/main/java/dev/onecoffeeplz/zamechalka/domain/repository/AudioRecordRepository.kt
@@ -2,7 +2,7 @@ package dev.onecoffeeplz.zamechalka.domain.repository
 
 import kotlinx.coroutines.CoroutineScope
 
-interface AudioRepository {
+interface AudioRecordRepository {
     suspend fun startRecording(scope: CoroutineScope): Result<Unit>
     suspend fun stopRecording(): Result<String>
     suspend fun deleteRecording(path: String): Result<Unit>
