+++ app/src/main/java/dev/onecoffeeplz/zamechalka/domain/usecase/impl/playback/PauseAudioPlayerUseCaseImpl.kt
@@ -1,7 +1,7 @@
-package dev.onecoffeeplz.zamechalka.domain.usecase.impl
+package dev.onecoffeeplz.zamechalka.domain.usecase.impl.playback
 
 import dev.onecoffeeplz.zamechalka.domain.repository.AudioPlayerRepository
-import dev.onecoffeeplz.zamechalka.domain.usecase.PauseAudioPlayerUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.playback.PauseAudioPlayerUseCase
 
 class PauseAudioPlayerUseCaseImpl(private val audioPlayerRepository: AudioPlayerRepository): PauseAudioPlayerUseCase {
     override fun invoke(): Result<Unit> {
