+++ app/src/main/java/dev/onecoffeeplz/zamechalka/domain/usecase/impl/playback/ReleaseAudioPlayerUseCaseImpl.kt
@@ -1,7 +1,7 @@
-package dev.onecoffeeplz.zamechalka.domain.usecase.impl
+package dev.onecoffeeplz.zamechalka.domain.usecase.impl.playback
 
 import dev.onecoffeeplz.zamechalka.domain.repository.AudioPlayerRepository
-import dev.onecoffeeplz.zamechalka.domain.usecase.ReleaseAudioPlayerUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.playback.ReleaseAudioPlayerUseCase
 
 class ReleaseAudioPlayerUseCaseImpl(private val audioPlayerRepository: AudioPlayerRepository): ReleaseAudioPlayerUseCase {
     override fun invoke(): Result<Unit> {
