+++ app/src/main/java/dev/onecoffeeplz/zamechalka/domain/usecase/impl/playback/StartAudioPlayerUseCaseImpl.kt
@@ -1,7 +1,7 @@
-package dev.onecoffeeplz.zamechalka.domain.usecase.impl
+package dev.onecoffeeplz.zamechalka.domain.usecase.impl.playback
 
 import dev.onecoffeeplz.zamechalka.domain.repository.AudioPlayerRepository
-import dev.onecoffeeplz.zamechalka.domain.usecase.StartAudioPlayerUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.playback.StartAudioPlayerUseCase
 
 class StartAudioPlayerUseCaseImpl(private val audioPlayerRepository: AudioPlayerRepository) :
     StartAudioPlayerUseCase {
