+++ app/src/main/java/dev/onecoffeeplz/zamechalka/domain/usecase/impl/recording/DeleteRecordingUseCaseImpl.kt
@@ -1,0 +1,11 @@
+package dev.onecoffeeplz.zamechalka.domain.usecase.impl.recording
+
+import dev.onecoffeeplz.zamechalka.domain.repository.AudioRecordRepository
+import dev.onecoffeeplz.zamechalka.domain.usecase.recording.DeleteRecordingUseCase
+
+class DeleteRecordingUseCaseImpl(private val audioRecordRepository: AudioRecordRepository) :
+    DeleteRecordingUseCase {
+    override suspend fun invoke(path: String): Result<Unit> {
+        return audioRecordRepository.deleteRecording(path)
+    }
+}
