+++ app/src/main/java/dev/onecoffeeplz/zamechalka/domain/usecase/impl/recording/StopRecordingUseCaseImpl.kt
@@ -1,0 +1,11 @@
+package dev.onecoffeeplz.zamechalka.domain.usecase.impl.recording
+
+import dev.onecoffeeplz.zamechalka.domain.repository.AudioRecordRepository
+import dev.onecoffeeplz.zamechalka.domain.usecase.recording.StopRecordingUseCase
+
+class StopRecordingUseCaseImpl(private val audioRecordRepository: AudioRecordRepository) :
+    StopRecordingUseCase {
+    override suspend fun invoke(): Result<String> {
+        return audioRecordRepository.stopRecording()
+    }
+}
