+++ app/src/main/java/dev/onecoffeeplz/zamechalka/domain/usecase/impl/view/CreateNoteUseCaseImpl.kt
@@ -1,8 +1,8 @@
-package dev.onecoffeeplz.zamechalka.domain.usecase.impl
+package dev.onecoffeeplz.zamechalka.domain.usecase.impl.view
 
 import dev.onecoffeeplz.zamechalka.domain.model.Note
 import dev.onecoffeeplz.zamechalka.domain.repository.NoteRepository
-import dev.onecoffeeplz.zamechalka.domain.usecase.CreateNoteUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.view.CreateNoteUseCase
 
 class CreateNoteUseCaseImpl(private val repository: NoteRepository) : CreateNoteUseCase {
     override suspend fun invoke(note: Note): Result<Unit> {
