+++ app/src/main/java/dev/onecoffeeplz/zamechalka/domain/usecase/impl/view/DeleteNoteUseCaseImpl.kt
@@ -1,8 +1,8 @@
-package dev.onecoffeeplz.zamechalka.domain.usecase.impl
+package dev.onecoffeeplz.zamechalka.domain.usecase.impl.view
 
 import dev.onecoffeeplz.zamechalka.domain.model.Note
 import dev.onecoffeeplz.zamechalka.domain.repository.NoteRepository
-import dev.onecoffeeplz.zamechalka.domain.usecase.DeleteNoteUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.view.DeleteNoteUseCase
 
 class DeleteNoteUseCaseImpl(private val repository: NoteRepository) : DeleteNoteUseCase {
     override suspend fun invoke(note: Note): Result<Unit> {
