+++ app/src/main/java/dev/onecoffeeplz/zamechalka/domain/usecase/impl/view/GetNoteByIdUseCaseImpl.kt
@@ -1,8 +1,8 @@
-package dev.onecoffeeplz.zamechalka.domain.usecase.impl
+package dev.onecoffeeplz.zamechalka.domain.usecase.impl.view
 
 import dev.onecoffeeplz.zamechalka.domain.model.Note
 import dev.onecoffeeplz.zamechalka.domain.repository.NoteRepository
-import dev.onecoffeeplz.zamechalka.domain.usecase.GetNoteByIdUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.view.GetNoteByIdUseCase
 import kotlinx.coroutines.flow.Flow
 
 class GetNoteByIdUseCaseImpl(private val repository: NoteRepository) : GetNoteByIdUseCase {
