+++ app/src/main/java/dev/onecoffeeplz/zamechalka/domain/usecase/impl/view/GetNotesUseCaseImpl.kt
@@ -1,8 +1,8 @@
-package dev.onecoffeeplz.zamechalka.domain.usecase.impl
+package dev.onecoffeeplz.zamechalka.domain.usecase.impl.view
 
 import dev.onecoffeeplz.zamechalka.domain.model.Note
 import dev.onecoffeeplz.zamechalka.domain.repository.NoteRepository
-import dev.onecoffeeplz.zamechalka.domain.usecase.GetNotesUseCase
+import dev.onecoffeeplz.zamechalka.domain.usecase.view.GetNotesUseCase
 import kotlinx.coroutines.flow.Flow
 
 class GetNotesUseCaseImpl(private val repository: NoteRepository): GetNotesUseCase {
