+++ app/src/main/java/dev/onecoffeeplz/zamechalka/domain/usecase/playback/ReleaseAudioPlayerUseCase.kt
@@ -1,4 +1,4 @@
-package dev.onecoffeeplz.zamechalka.domain.usecase
+package dev.onecoffeeplz.zamechalka.domain.usecase.playback
 
 interface ReleaseAudioPlayerUseCase {
     operator fun invoke(): Result<Unit>
