+++ app/src/main/java/dev/onecoffeeplz/zamechalka/domain/usecase/recording/DeleteRecordingUseCase.kt
@@ -1,4 +1,4 @@
-package dev.onecoffeeplz.zamechalka.domain.usecase
+package dev.onecoffeeplz.zamechalka.domain.usecase.recording
 
 interface DeleteRecordingUseCase {
     suspend operator fun invoke(path: String): Result<Unit>
