+++ app/src/main/java/dev/onecoffeeplz/zamechalka/domain/usecase/recording/StopRecordingUseCase.kt
@@ -1,4 +1,4 @@
-package dev.onecoffeeplz.zamechalka.domain.usecase
+package dev.onecoffeeplz.zamechalka.domain.usecase.recording
 
 interface StopRecordingUseCase {
     suspend operator fun invoke(): Result<String>
